Reimplement log buttons using CSS.
The CASA 5.X version of jupyter-casa used Javascript for log buttons,
unfortunately, this code needed to be replaced because JupyterLab
removed support for inline Javascript.

Our first attempt used ipywidgets. However, this was not as
straightforward as was hoped. Because the ipywidgets are controlled
by the kernel, the log buttons became unresponsive while cells
were evaluated. Also there were issues in saving widget contents to
the notebooks.