Bump OS to ubuntu 22.04
The new ubuntu version allows us to use the lastest kern suite,
with goodies like aoflagger, carta, wsclean, ..

Unfortunately, this does mean loosing the almatasks package as
this package is only available for python 3.8 at the moment.
As far as I can tell almatasks only contains one task (wvrgcal),
and this doesn't seem to be needed for end-users of alma data.